Move expression based validations into ExpressionsJavaValidator (#1247)
