+++ plugins/org.yakindu.base.expressions/src/org/yakindu/base/expressions/validation/ExpressionsJavaValidator.java
@@ -12,16 +12,28 @@
 package org.yakindu.base.expressions.validation;
 
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
 
 import org.eclipse.emf.common.util.EList;
+import org.eclipse.emf.ecore.EClass;
+import org.eclipse.emf.ecore.EObject;
 import org.eclipse.xtext.validation.Check;
+import org.eclipse.xtext.validation.CheckType;
 import org.yakindu.base.expressions.expressions.Argument;
 import org.yakindu.base.expressions.expressions.ArgumentExpression;
+import org.yakindu.base.expressions.expressions.AssignmentExpression;
+import org.yakindu.base.expressions.expressions.ElementReferenceExpression;
 import org.yakindu.base.expressions.expressions.Expression;
+import org.yakindu.base.expressions.expressions.ExpressionsPackage;
+import org.yakindu.base.expressions.expressions.FeatureCall;
+import org.yakindu.base.types.AnnotatableElement;
+import org.yakindu.base.types.Annotation;
 import org.yakindu.base.types.ComplexType;
 import org.yakindu.base.types.GenericElement;
+import org.yakindu.base.types.Operation;
 import org.yakindu.base.types.Parameter;
+import org.yakindu.base.types.Property;
 import org.yakindu.base.types.Type;
 import org.yakindu.base.types.TypeParameter;
 import org.yakindu.base.types.TypeSpecifier;
@@ -30,6 +42,8 @@ import org.yakindu.base.types.inferrer.ITypeSystemInferrer;
 import org.yakindu.base.types.typesystem.ITypeSystem;
 import org.yakindu.base.types.validation.IValidationIssueAcceptor;
 
+import com.google.common.base.Predicate;
+import com.google.common.collect.Iterables;
 import com.google.common.collect.Sets;
 import com.google.inject.Inject;
 
@@ -61,6 +75,21 @@ public class ExpressionsJavaValidator extends org.yakindu.base.expressions.valid
 
 	public static final String ERROR_DUPLICATE_PARAMETER_ASSIGNMENT_CODE = "ErrorDuplicateParameterAssignment";
 	public static final String ERROR_DUPLICATE_PARAMETER_ASSIGNMENT_MSG = "Duplicate assignment to parameter '%s'.";
+	
+	public static final String ERROR_ASSIGNMENT_TO_CONST_MSG = "Assignment to constant not allowed.";
+	public static final String ERROR_ASSIGNMENT_TO_CONST_CODE = "AssignmentToConst";
+	
+	public static final String ERROR_LEFT_HAND_ASSIGNMENT_MSG = "The left-hand side of an assignment must be a variable.";
+	public static final String ERROR_LEFT_HAND_ASSIGNMENT_CODE = "LeftHandAssignment";
+	
+	public static final String ERROR_WRONG_NUMBER_OF_ARGUMENTS_MSG = "Wrong number of arguments, expected %s .";
+	public static final String ERROR_WRONG_NUMBER_OF_ARGUMENTS_CODE = "WrongNrOfArgs";
+	
+	public static final String ERROR_VAR_ARGS_LAST_CODE = "VarArgsMustBeLast";
+	public static final String ERROR_VAR_ARGS_LAST_MSG = "The variable argument type must be the last argument.";
+	
+	public static final String ERROR_WRONG_ANNOTATION_TARGET_CODE = "WrongAnnotationTarget";
+	public static final String ERROR_WRONG_ANNOTATION_TARGET_MSG = "Annotation '%s' can not be applied on %s ";
 
 	@Inject
 	private GenericsPrettyPrinter printer;
@@ -191,6 +220,93 @@ public class ExpressionsJavaValidator extends org.yakindu.base.expressions.valid
 					break;
 				}
 				assignedParameters.add(argument.getParameter());
+			}
+		}
+	}
+	
+	@Check(CheckType.FAST)
+	public void checkAssignmentToFinalVariable(AssignmentExpression exp) {
+		Expression varRef = exp.getVarRef();
+		EObject referencedObject = null;
+		if (varRef instanceof FeatureCall)
+			referencedObject = ((FeatureCall) varRef).getFeature();
+		else if (varRef instanceof ElementReferenceExpression)
+			referencedObject = ((ElementReferenceExpression) varRef).getReference();
+		if (referencedObject instanceof Property) {
+			if (((Property) referencedObject).isConst()) {
+				error(ERROR_ASSIGNMENT_TO_CONST_MSG, ExpressionsPackage.Literals.ASSIGNMENT_EXPRESSION__VAR_REF, ERROR_ASSIGNMENT_TO_CONST_CODE);
+			}
+		}
+	}
+	
+	@Check(CheckType.FAST)
+	public void checkLeftHandAssignment(final AssignmentExpression expression) {
+		Expression varRef = expression.getVarRef();
+		if (varRef instanceof FeatureCall) {
+			EObject referencedObject = ((FeatureCall) varRef).getFeature();
+			if (!(referencedObject instanceof Property)) {
+				error(ERROR_LEFT_HAND_ASSIGNMENT_MSG, ExpressionsPackage.Literals.ASSIGNMENT_EXPRESSION__VAR_REF, ERROR_LEFT_HAND_ASSIGNMENT_CODE);
+			}
+		} else if (varRef instanceof ElementReferenceExpression) {
+			EObject referencedObject = ((ElementReferenceExpression) varRef).getReference();
+			if (!(referencedObject instanceof Property)) {
+				error(ERROR_LEFT_HAND_ASSIGNMENT_MSG, ExpressionsPackage.Literals.ASSIGNMENT_EXPRESSION__VAR_REF, ERROR_LEFT_HAND_ASSIGNMENT_CODE);
+			}
+
+		} else {
+			error(ERROR_LEFT_HAND_ASSIGNMENT_MSG, ExpressionsPackage.Literals.ASSIGNMENT_EXPRESSION__VAR_REF, ERROR_LEFT_HAND_ASSIGNMENT_CODE);
+		}
+	}
+
+	@Check(CheckType.FAST)
+	public void checkOperationArguments_FeatureCall(final FeatureCall call) {
+		if (call.getFeature() instanceof Operation) {
+			Operation operation = (Operation) call.getFeature();
+			assertOperationArguments(operation, call.getExpressions());
+		}
+	}
+
+	@Check(CheckType.FAST)
+	public void checkOperationArguments_TypedElementReferenceExpression(final ElementReferenceExpression call) {
+		if (call.getReference() instanceof Operation) {
+			Operation operation = (Operation) call.getReference();
+			assertOperationArguments(operation, call.getExpressions());
+		}
+	}
+
+	protected void assertOperationArguments(Operation operation, List<Expression> args) {
+		EList<Parameter> parameters = operation.getParameters();
+		if ((operation.isVariadic() && operation.getVarArgIndex() > args.size())
+				|| !operation.isVariadic() && parameters.size() != args.size()) {
+			error(String.format(ERROR_WRONG_NUMBER_OF_ARGUMENTS_MSG, parameters), null, ERROR_WRONG_NUMBER_OF_ARGUMENTS_CODE);
+		}
+	}
+	
+	@Check(CheckType.FAST)
+	public void checkVarArgParameterIsLast(Operation operation) {
+		if (operation.isVariadic() && operation.getVarArgIndex() != operation.getParameters().size() - 1) {
+			error(ERROR_VAR_ARGS_LAST_MSG, operation.getParameters().get(operation.getVarArgIndex()), null,
+					ERROR_VAR_ARGS_LAST_CODE);
+		}
+	}
+	
+	@Check(CheckType.FAST)
+	public void checkAnnotationTarget(final AnnotatableElement element) {
+		EList<Annotation> annotations = element.getAnnotations();
+		for (Annotation annotation : annotations) {
+			EList<EObject> targets = annotation.getType().getTargets();
+			if(targets.isEmpty())
+				continue;
+			boolean found = Iterables.any(targets, new Predicate<EObject>() {
+				@Override
+				public boolean apply(EObject input) {
+					return ((EClass) input).isInstance(element);
+				}
+			});
+			if (!found) {
+				error(String.format(ERROR_WRONG_ANNOTATION_TARGET_MSG, annotation.getType().getName(),
+						element.eClass()), null,
+						element.getAnnotations().indexOf(annotation), ERROR_WRONG_ANNOTATION_TARGET_CODE);
 			}
 		}
 	}
